requirements.txt
asgiref==3.2.10
certifi==2021.5.30
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1
django-filter==2.4.0
djangorestframework==3.12.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==21.0
Pillow==7.1.0
pkg-resources==0.0.0
psycopg2-binary==2.8.6
PyJWT==1.7.1
pyparsing==2.4.7
python-decouple==3.4
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
transliterate==1.10.2
uritemplate==3.0.1
urllib3==1.26.6
